fix: move types to analytics-types
