+++ packages/analytics-browser/src/typings/browser-snippet.d.ts
@@ -1,16 +1,4 @@
-import { Result } from '@amplitude/analytics-types';
-
-interface ProxyItem {
-  name: string;
-  args: any[];
-  resolve?: (promise: Promise<Result>) => void;
-}
-
-export type QueueProxy = Array<ProxyItem>;
-
-export interface InstanceProxy {
-  _q: QueueProxy;
-}
+import { InstanceProxy } from '@amplitude/analytics-types';
 
 declare global {
   // globalThis only includes `var` declarations
